requirements.txt
streamlit==1.10.0
protobuf==3.19.1
click<=8.0.4
jpholiday==0.1.8
google-api-core==2.3.2
google-api-python-client
google-auth==2.1.0
google-auth-httplib2
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
gspread==4.0.1
oauth2client==4.1.3
PyDrive==1.3.1